shifted rx letter head details from doctor to user_clinic table. Resolved issue for doctor. Applied Patch for other roles as of now.
