+++ src/doctor/entity/doctor.entity.ts
@@ -40,26 +40,6 @@ export class Doctor {
     endTime: Date; // Format as HH:MM
   }[];
 
-  @Column({ type: 'text', nullable: true })
-  headerImage: Buffer | '';
-
-  @Column({ type: 'text', nullable: true })
-  footerText: string | null;
-
-  @Column({
-    type: 'json',
-    nullable: true,
-  })
-  padding: {
-    paddingTop: number | null;
-    paddingLeft: number | null;
-    paddingBottom: number | null;
-    paddingRight: number | null;
-  };
-
-  @Column({ default: true })
-  usesOwnLetterPad: boolean;
-
   @OneToOne(() => User, (user) => user.doctor, {
     nullable: true,
     onDelete: 'CASCADE',
