+++ src/user_clinic/dto/create_user_clinic.dto.ts
@@ -1,4 +1,4 @@
-import { IsEnum, IsInt, IsNotEmpty, IsString } from 'class-validator';
+import { IsBoolean, IsEnum, IsInt, IsNotEmpty, IsOptional, IsString, MaxLength } from 'class-validator';
 import { UserRole } from '../entity/user_clinic.entity';
 
 export class CreateUserClinicDto {
@@ -13,4 +13,24 @@ export class CreateUserClinicDto {
   @IsNotEmpty()
   @IsEnum(UserRole, { each: true })
   role: UserRole[];
+
+  @IsOptional()
+  headerImage?: Buffer;
+
+  @IsOptional()
+  @IsString()
+  @MaxLength(100)
+  footerText?: string;
+
+  @IsOptional()
+  padding?: {
+    paddingTop?: number | null;
+    paddingLeft?: number | null;
+    paddingBottom?: number | null;
+    paddingRight?: number | null;
+  };
+
+  @IsOptional()
+  @IsBoolean()
+  usesOwnLetterPad?: boolean;
 }
